Connack handlers

+++ src/client.h
@@ -1,9 +1,9 @@
-
 #include <stddef.h>
+#include <sys/types.h> // Include sys/types.h for ssize_t
 
 #ifndef CLIENT_H
 #define CLIENT_H
 
-extern int bytes_rw;
+extern ssize_t bytes_rw;
 
 #endif
